+++ python-lib/ontology_tagging/__init__.py
@@ -1,0 +1,6 @@
+from .ontology_tagger import Tagger
+from .dku_plugin_config_loading import DkuConfigLoading,DkuConfigLoadingOntologyTagging
+from .formatter_instanciator import FormatterInstanciator
+from .ontology_tagging_formatting import Formatter
+from .sentence_splitter import SentenceSplitter
+from .spacy_tokenizer import MultilingualTokenizer
